+++ src/pages/book/detail/index.tsx
@@ -1,28 +1,16 @@
 import { dehydrate, QueryClient } from '@tanstack/react-query';
-import axios from 'axios';
 import { GetServerSideProps } from 'next';
 import React from 'react';
 
+import { fetchDummy } from '@/api/dummy';
 import BookDescription from '@/components/book/detail/BookDescription';
 import BookList from '@/components/book/detail/BookList';
 import Buttons from '@/components/book/detail/Buttons';
 
-export interface BookListProps {
-  books: {
-    id: string;
-    title: string;
-  }[];
-}
-
-export interface BooksProps {
-  id: string;
-  title: string;
-  description: string;
-}
-
 const BookDetailPage = () => {
   return (
     <div className='h-screen p-5 border-2 border-red-500'>
+      {/* 디자인이 나오면 정확한 수치로 Layout컴포넌트 만들어 _app.tsx에 감싸주기 */}
       <main className='h-full flex flex-col gap-5'>
         <BookDescription />
         <Buttons />
@@ -30,21 +18,6 @@ const BookDetailPage = () => {
       </main>
     </div>
   );
-};
-
-export const fetchDummy = async () => {
-  const resposne = await axios.get(
-    'https://dummy-cca81-default-rtdb.asia-southeast1.firebasedatabase.app/books.json',
-  );
-  const formatBooks: BooksProps[] = [];
-  for (const key in resposne.data) {
-    formatBooks.push({
-      id: key,
-      title: resposne.data[key].title,
-      description: resposne.data[key].description,
-    });
-  }
-  return formatBooks;
 };
 
 export const getServerSideProps: GetServerSideProps = async () => {
